% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_kriging_vgm.R
\name{fit_kriging_vgm}
\alias{fit_kriging_vgm}
\title{fit kriging vgm model}
\usage{
fit_kriging_vgm(s, range = NULL, show)
}
\arguments{
\item{s}{SpatialPointsDataFrame, which is used to be interpolated}

\item{range}{vgm parameter, distance range, in the unit of m}

\item{show}{boolean, whether plot output?}
}
\description{
Two simple vgm model were used:
(1) Exp
(2) Lin: if \code{Exp} model failed to converge, then \code{Lin} used.
}
