% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_transform.R
\docType{data}
\name{prj_g_CGCS2000}
\alias{prj_g_CGCS2000}
\title{CGCS2000 Geographic coordinate system}
\format{
An object of class \code{CRS} of length 1.
}
\usage{
prj_g_CGCS2000
}
\value{
CRS arguments:
+init = epsg:4490 + proj = longlat + ellps = GRS80 + no_defs
}
\description{
CGCS2000 Geographic coordinate system
}
\details{
CGCS2000_3_Degree_GK_CM_90E
\itemize{
\item WKID: 4539 Authority: EPSG
\item \code{Projection}: Gauss_Kruger
\item \code{False_Easting}: 500000.0
\item \code{False_Northing}: 0.0
\item \code{Central_Meridian}: 90.0
\item \code{Scale_Factor}: 1.0
\item \code{Latitude_Of_Origin}: 0.0
\item \verb{Linear Unit}: Meter (1.0)
}
}
\examples{
prj_g_CGCS2000
}
\keyword{datasets}
