% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_transform.R
\docType{data}
\name{prj_p_CGCS2000}
\alias{prj_p_CGCS2000}
\title{CGCS2000 Projected coordinate system}
\format{
An object of class \code{CRS} of length 1.
}
\usage{
prj_p_CGCS2000
}
\value{
CRS arguments:
+init=epsg:4518 +proj=tmerc +lat_0=0 +lon_0=90 +k=1 +x_0=30500000 +y_0=0 +ellps=GRS80 +units=m +no_defs
}
\description{
Projected coordinate system is defined on a flat, two-dimensional surface
}
\details{
\code{CGCS2000_3_Degree_GK_Zone_30}
\itemize{
\item \code{WKID}: 4518 Authority: EPSG
\item \code{Projection}: Gauss_Kruger
\item \code{False_Easting}: 30500000.0
\item \code{False_Northing}: 0.0
\item \code{Central_Meridian}: 90.0
\item \code{Scale_Factor}: 1.0
\item \code{Latitude_Of_Origin}: 0.0
\item \verb{Linear Unit}: Meter (1.0)
}
}
\examples{
prj_p_CGCS2000
}
\keyword{datasets}
