% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_transform.R
\name{to_CGCS2000}
\alias{to_CGCS2000}
\title{Convert to CGCS2000 Projected coordinate system}
\usage{
to_CGCS2000(x)
}
\arguments{
\item{x}{Spatial* object}
}
\description{
Convert to CGCS2000 Projected coordinate system
}
